+++ lib/glimmer/dsl/swt/rgba_expression.rb
@@ -1,0 +1,12 @@
+require 'glimmer/dsl/swt/color_expression'
+require 'glimmer/swt/color_proxy'
+require 'glimmer/swt/display_proxy'
+
+module Glimmer
+  module DSL
+    module SWT
+      class RgbaExpression < ColorExpression
+      end
+    end
+  end
+end