+++ lib/glimmer/dsl/top_level_expression.rb
@@ -1,0 +1,7 @@
+module Glimmer
+  module DSL
+    # Mixin that represents expressions that are always at the top with parent nil
+    module TopLevelExpression
+    end
+  end
+end