+++ lib/glimmer/dsl/xml/dsl.rb
@@ -4,7 +4,6 @@ Dir[File.expand_path('../*_expression.rb', __FILE__)].each {|f| require f}
 Glimmer::DSL::Engine.add_dynamic_expressions(
   Glimmer::DSL::XML,
   %w[
-    xml_name_space
     xml_text
     xml_tag
     html_tag